pyproject.toml
[project]
name = "serial-cli"
description = "Command line interface for serial communication"
authors = [{ name = "André Escorel", email = "gustavo.escorel@gmail.com" }]
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
dynamic = ["version", "readme"]
dependencies = [
    "click>=8.2.1",
    "prompt-toolkit>=3.0.52",
    "pyserial>=3.5",
    "rich>=14.1.0",
    "rich-click>=1.8.9",
]

[project.urls]
Homepage = "https://github.com/commonProgrammer/serial-cli"
repository = "https://github.com/commonProgrammer/serial-cli"

[project.scripts]
serial-cli = "serial_cli.cli:main"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=8.37.0",
    "pytest-watch>=4.2.0",
    "taskipy>=1.14.1",
]

[project.optional-dependencies]
test = [
    "coverage>=7.6.4",
    "pytest>=8.4.1",
    "pytest-cov>=5.0.0",
    "pytest-forked>=1.6.0",
    "types-setuptools>=75.2.0.20241025",
]

[build-system]
requires = ["setuptools>=80.9.0", "setuptools_scm[toml]>=6.2"]
build-backend = 'setuptools.build_meta'

[tool.setuptools]
packages = ["serial_cli"]

[tool.setuptools.dynamic]
readme = { file = ["README.md"] }

[tool.setuptools_scm]
write_to = "VERSION.txt"


[tool.ruff]
exclude = [".venv"]
line-length = 79
indent-width = 4
target-version = "py312"
lint.extend-select = ["I"]
lint.extend-ignore = ["E203", "E701"]

[tool.taskipy.tasks]
lint = { help = "Run ruff Linter", cmd = "uvx ruff check ." }
fmt = { help = "Run ruff formatter", cmd = "uvx ruff format .; uvx ruff check --select I --fix ." }
clean = { help = "Clean environment", cmd = """\
#! /usr/bin/env bash
find ./ -name '*.pyc' -exec rm -f {} \\;
find ./ -name '__pycache__' -exec rm -rf {} \\;
find ./ -name 'Thumbs.db' -exec rm -f {} \\;
find ./ -name '*~' -exec rm -f {} \\;
rm -rf {.cache,.pytest_cache,.mypy_cache,htmlcov,docs/_build}""" }
